Constant HTMLY condition for Templates
The theme folder and the php templates are accessible direct from the
webbrowser. To avoid output of their content, this commit adds a
codition to each template. This condition tests if the constant "HTMLY"
is set. If it's not set, the template outputs nothing. ("Silence is
golden...")

This update is mainly to suppress the template's output but also to
avoid output of php errors, if it's opened directly from the webbrowser.

Update in index.php

define('HTMLY', true);

Update in each template php file.

<?php if (!defined('HTMLY')) exit(); ?>

Example for a direct opened template:
https://www.htmly.com/themes/default/profile.html.php

After this update, this url outputs a blank page only.