Replace RSS description value with Meta Description instead of Body
The Meta Description of a post makes more sense for the RSS feed since it's either a user generated description of the post or automatically pulled from the post body. A user can leave the Meta Description blank to use the post body for the RSS feed (which is basically the current RSS method).